refactor(ResultView) : OutputView -> ResultView 로 이름 변경 및 플레이어의 카드 출력 기능 추가
